fix: use folder structure and align styles to index
