+++ src/components/Feed/styles/index.jsx
@@ -1,23 +1,20 @@
 import styled from 'styled-components';
 
 const Wrapper = styled.div`
+  ${({ theme }) => theme.mixins.flexbox('column', 'initial', 'initial', '43px')};
+
   width: 100%;
-  display: flex;
-  flex-direction: column;
   padding: 78px 0;
-  gap: 43px;
   height: 100%;
   overflow-y: auto;
 `;
 
 const Header = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
+  ${({ theme }) => theme.mixins.flexbox('row', 'center', 'flex-start', 'initial')};
 `;
 
 const Title = styled.h1`
-  font-family: Oswald;
+  font-family: ${({ theme }) => theme.fonts.logotype};
   font-size: 43px;
   font-weight: 700;
   line-height: 64px;
@@ -25,27 +22,22 @@ const Title = styled.h1`
 `;
 
 const UserThumbnail = styled.img`
-  width: 50px;
+  ${({ theme }) => theme.mixins.flexbox('row', 'center', 'center', 'initial')};
+  margin-right: 20px;
   height: 50px;
+  width: 50px;
+
+  object-fit: cover;
   border-radius: 50%;
   background-color: #f5f5f5;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 20px;
-  object-fit: cover;
 `;
 
 const Content = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
+  ${({ theme }) => theme.mixins.flexbox('column', 'initial', 'initial', '30px')};
 `;
 
 const Posts = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
+  ${({ theme }) => theme.mixins.flexbox('column', 'initial', 'initial', '16px')};
 `;
 
 export { Wrapper, Header, Title, UserThumbnail, Content, Posts };
